Add back cach logic. added tablet (nexus 9) layout for image viewer layout
Image viewer for tablet still needs tweaking but this is a good first
draft.
Added back the caching logic to reduce repeat downloads of previously
seen files.  This slipped through.  I had added the clear() to do some
testing and forgot to remove it..  I decided to go ahead and clear the
cache after about 1000 pictures are viewed.